# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: RDV.BI API
  version: 0.1.0
  summary: Appointment booking, queue, and institution administration API
  description: |
    OpenAPI contract for the current RDV.BI backend.

    Notes:
    - Public endpoints live under `/api/public`.
    - Authentication endpoints live under `/api/auth`.
    - Admin endpoints require a bearer JWT issued by `/api/auth/login`.
    - `POST /api/auth/bootstrap-admin` exists only for bootstrap and should be removed or disabled in production.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/admin/appointments:
    get:
      tags: [Admin Appointments]
      summary: List appointments or fetch one by filters
      security:
        - BearerAuth: []
      parameters:
        - name: institutionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: appointmentDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: customerPhone
          in: query
          required: false
          schema:
            type: string
        - name: id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Appointment list or a single appointment when `id` is used
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Appointment'
                  - $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/appointments/{id}:
    put:
      tags: [Admin Appointments]
      summary: Update an appointment
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdateRequest'
      responses:
        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example:
        error: Institution not found
...
    AppointmentUpdateRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        appointment_date:
          type: string
          format: date
        appointment_time:
          type: string
          format: time
        time_slot_id:
          type: string
          format: uuid
        ticket_number:
          type: string
        customer_email:
          type: string
          format: email
        customer_phone:
          type: string
        counter_number:
          type: string
        called_time:
          type: string
          format: date-time
        completed_time:
          type: string
          format: date-time
...
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerPhone:
          type: string
        customerEmail:
          type: string
          format: email
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
          format: time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        ticketNumber:
          type: string
        otpVerified:
          type: boolean
        counterNumber:
          type: string
        calledTime:
          type: string
          format: date-time
        completedTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - appointmentDate
        - appointmentTime
        - status
...
    AppointmentStatus:
      type: string
      enum: [PENDING, CONFIRMED, WAITING, IN_PROGRESS, COMPLETED, CANCELLED, NO_SHOW]
...